db to linear

+++ pttw/signal_to_noise.go
@@ -1,6 +1,6 @@
 package pttw
 
-func AddSignalToNoise(signal []float64, noise []float64, snr int64) []float64 {
+func AddSignalToNoise(signal []float64, noise []float64, snr float64) []float64 {
 	sum := []float64{}
 	for i := 0; i < len(signal); i++ {
 		sum = append(sum, signal[i] + noise[i])
